Check if the Host is powered off before probing the console
Before we spend time trying to detect what system the state is in by
writing to the host console we should check if the host is powered on
or not.

I don't think I've seen a BMC give a false-off so this should be
reliable and save a bit of time on the probing. This isn't critical for
CI runs, but it's nice when using op-test manually from the command
line.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>