Remove timestamp prefixes from logfile
We already put them in a timestamped directory. There's not much point
in repeating it on the individual log files and as the logs are rotated
the timestamps drift which makes then annoying to tab complete.

If someone really wants to keep the timestamps then we could change the
prefix to a suffix I guess.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>