Adjust OpenBMC PowerOff and global SniffTest
In early OpenBMC targets, we were told to issue both immediate (chassis)
as well as soft (host) poweroff in our sys_power_off() path. This is no
longer the recommendation, so adjust to be consistent with IPMI and LPAR
methods that issue an immediate power off.

Additionally, we're also been told that 30 seconds might not be enough
even for an immediate PowerOff in some (upcoming?) systems, so adjust
the SniffTest logic so we wait_for_standby() before issuing a PowerOn.

This should fix some CI scenarios where the BMC was executing a soft
power off (even if we're issuing both immediate and soft) and the 30
second delay was not sufficient before we could issue a poweron -
intermintently failing the test.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>