OpTestOpenBMC: make query_vpnor check for pflash
While investigating an issue where op-test would fail to flash a OpenBMC
system with a new PNOR, it became clear that recent OpenBMC systems can
have 'empty' inventory lists for host firmware, which was the condition
being used so far to make a determination if the system was capable of
VPNOR or not.

Further investigation with the OpenBMC team revealed that most of the
OpenBMC builds today support VPNOR, and the only reliable way to check
is the presence of 'pflash' bundled as part of BMC root filesystem.

If 'pflash' is present, means that OpenBMC doesn't support VPNOR.
Otherwise, it should.

Fixes said scenarios described above where op-test would fail to flash
OpenBMC systems with an empty inventory list of host images.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>