testcases.Console.Control[C|Z]: Increase timeout
Due to an OpenBMC Bug we're intermittently hitting a timeout on
ControlC/ControlZ Console tests. Work around this by doubling the
timeout. On my tests, each iteration takes between 60-140 seconds, so
the 5 iterations are taking a good 10 minutes to complete, also
generating *a lot* of debug logs.

But before the OpenBMC bug is fixes, it's better than failing and
re-trying the test.

Once this is fixed, we should put this back to a sane default (say 10
seconds?)

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>