Travis CI: use Witherspoon instead of Palmetto
Palmetto (or any other Power8 targets) is no longer supported in
upstream op-build, so move to use Witherspoon pnors instead.

Also move to use Cedric's powernv-5.1 Qemu branch as it is the only
one supporting both powernv(9) as well as Skiboot lid override.
This requires a few changes in the PHB assignment, as we now
have 6 PHBs instead of just 3.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>