Defined a command-line argument and function calls in OpTestLPM.py
Defined a command-line argument (--collect-pre-post-test-logs)
which needs to be passed if the user wishes to collect logs
before the test begins and after the test ends.

In OpTestLPM.py, calling functions to collect logs and
parse dmesg for errors at different stages:
- Before test begins
- After test ends
- Within the test - pre-forward-lpm, post-forward-lpm, post-backward-lpm
- When test fails

Also, calling function to clear dmesg at the beginning of test.

Defined some additional attributes that need to be passed for
collecting VIOS and HMC logs.
- vios_ip, vios_username, vios_password
- hmc_hscpe_password
And in the case of Cross LPM:
- remote_vios_ip, remote_vios_username, remote_vios_password
- remote_hmc_hscpe_password

Signed-off-by: rashijhawar <rashi@linux.vnet.ibm.com>