Modified the sed commands (kdump)
For kdump using SSH,
	- commenting existing lines in /etc/kdump.conf that start with 'nfs'
	- commenting existing lines in /etc/kdump.conf that start with
	  'ssh ', 'sshkey', and 'path', and adding new lines for each

For kdump using NFS,
	- commenting existing lines in /etc/kdump.conf that start with 'ssh '
	  and 'sshkey'
	- commenting existing lines in /etc/kdump.conf that start with
          'nfs', and 'path', and adding new lines for each

This updated code takes care of adding these params in /etc/kdump.conf
even if they are not already present.

And it also matches more generically. For instance, initially the code
would search for 'path /var/crash' in /etc/kdump.conf and replace it.
If instead, "path /" was present in /etc/kdump.conf, there would be no match.

- Replaced the sed commands at the end of the KernelCrash_KdumpNFS test
with setup_test(), so as to revert the config file to what it was before
the test started (like done in KernelCrash_KdumpSSH).

Signed-off-by: rashijhawar <rashi@linux.vnet.ibm.com>