testcases/InstallUpstreamKernel: boot order(grub2) fixes
commit 258fb8545f1f: Use grub2-set-default inplace of grubby
switched to using grub2-set-default instead of grubby since
that command is available on most of distributions.

While that command works most of the time on certain instances
inspite of running that command the kernel boot order is not
set as required. This causes system to boot into a wrong kernel.

Add a check for distro version/release and accordingly use
either grub2-set-default or grubby command

Also make modules step is not required so remove it.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>