op-test/common: Skip fsp_get_console call
OpTestLPARSystem iniitalization includes a call to fsp_get_console.
This function call disables firewall, opens a connection to FSP and
records hostname. This call isn't necessary for LPAR based system
and possibly was inherited from OpTestFSPSystem class.

Don't call this function so that it does not alter the FSP
configuration.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>