Test to verify energyd service functionality - v2
This test covers below scenarios:
Starts energyd service and checks if it is active.
Verifies if the number of online cores reduces to 2 when service is active in powersave mode.
This test verifies that all cores are online in max performance mode even if service is active.
This test verifies that all cores come online when workload is started even in powersave mode.

Signed-off-by: Pavithra <pavrampu@linux.vnet.ibm.com>