OpTestHMC/OpTestKexec : Enable secureboot pre-requisites
OpTestHMC is added with a new class "OpSecureBootUtilities"
which will incorporate all the common required secureboot
related APIs as pre-requisites. This commit addresses two
API additions as of now -
1. check_kernel_signature()
2. check_os_level_secureboot_state()

OpTestKexec is changed to handle the returned values from
the above two APIs which would be required further to
handle secureboot enabled kexec operations.

Signed-off-by: Rohan Deshpande <rohan_d@linux.ibm.com>