OpTestKexec: simplify setUp function
Refactor the `setUp` function of `OpTestKexec` by relocating the code
related to finding the kexec vmlinux and initrd image based on the
distro to a separate function.

Also get rid of distro instance variable because it was only used in
setUp function.

This improvement enhances the code's readability without introducing
any functional changes.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>