OpTestKexec: revamp kexec test cases
There are two big challenges with the current kexec tests:

1. Not all kexec test cases get executed, but still show "pass" for all.
2. There is no way for the user to execute individual test cases.

The reason not all test cases are getting executed is that after the
first kexec test, the kexec kernel boots and the rest of the commands
are executed without logging into the system.

Additionally, since all the test functions are placed one after another
in a single `runTest` function, it restricts the user from running
specific tests even if they need to.

The changes done to solve the above problem are as follows:

To fix the first issue the system state is set to OpSystemState.UNKNOWN
and then OpSystemState.OS after every test to bring the system into
consistent state before performing next test. This is helpful for test
except `kexec in loop test

All test functions are now prefixed with test_ that enabled unittest
python module to identify tests in OpTestKexec class. With this user
can run kexec test to run individually.

The changes made to address the aforementioned problem are as follows:

To rectify the first issue, the system state is set to
OpSystemState.UNKNOWN, and then to OpSystemState.OS after every test, in
order to bring the system into a consistent state before performing the
next test. This is helpful for all tests except the 'kexec in loop'
test. For 'kexec in loop' test the execute_kexec_cmd function should be
bring the system in consistent state without rebooting the system. Which
is done by closing the raw_pty console so that when console is take for
next kexec login will be performed once again.

All test functions are now prefixed with 'test_', enabling the unittest
Python module to identify tests in the OpTestKexec class. With this,
users can run the kexec test individually.

Additionally, a few helper functions have been introduced to modularize
the kexec test and reduce code redundancy.

The function kexec_suite has been removed because it is not currently
being used. However, the patch titled "OpTestKexec: add test suite for
kexec" adds the kexec suite and includes it as part of the op-test suite
list.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>