OpTestKexec: add test suite for kexec
The kexec suite executes the following test cases in the same order as
they are listed below:

1. kexec load and unload
2. kexec load and exec into the new kernel
3. kexec using the force option
4. kexec using single step
5. kexec in a loop

The KexecSuite class is updated to call the new function introduced for
invoking the kexec suite function defined in the OpTestKexec.py file.
Additionally, the kexec suite is added to the op-test suite list, so
user can execute the kexec suite using the --run-suite option.

For example:
$ ./op-test -c my-system.config --run-suite kexec-suite

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>