feat: allow op-test usage from anywhere
Currently, if a symlink to 'op-test' is created somewhere (such as in
/usr/local/bin/), it fails to start with following error:

```
$ cd /usr/local/bin
$ ln -sv PATH_TO_op-test/op-test
$
$ op-test --help

Traceback (most recent call last):
  File "/tmp/./op-test2", line 125, in <module>
    OpTestConfiguration.conf = OpTestConfiguration.OpTestConfiguration()
                               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tmp/op-test/OpTestConfiguration.py", line 492, in __init__
    for dirname in (next(os.walk(os.path.join(self.basedir, 'addons')))[1]):
                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
StopIteration
```

This is because, `self.basedir` is pointing to the current working
directory or path of the symlink (sys.argv[0]), and it didn't find
self.basedir/'addons' folder.

Solve this by ensuring `self.basedir` points to the directory containing
the op-test script.

This enables us to install op-test symlink in somewhere like
`/usr/local/bin`, and thus being able to run it from anywhere.

Signed-off-by: Aditya Gupta <adityag@linux.ibm.com>