Added support to configure the multiple lpars-v5
This enhancement introduces support for configuring multiple lpars, such as lpar1, lpar2, and so on, especially when they are in shared mode.

To configure the multiple lpars user need to pass below two parameters, lpar_config=multi
lpar_list=lpar1, lpar2 ...etc
In this case, users are required to explicitly specify these two parameters in the .conf file, and it's important to ensure that all lpars have unique passwords.

Signed-off-by: Samir Mulani <samir@linux.vnet.ibm.com>