Added support to take max_proc_units as input from user.
In the previous implementation, we did not support taking the max_proc_units value from user input. Instead, we first calculated the stealable resources and used that value, with a default setting of max_proc_units=2.
Now, we have added the ability to take max_proc_units as input from a configuration file. If the user does not provide this value, we fall back to the stealable resource count.

Signed-off-by: Samir Mulani <Samir.Mulani@ibm.com>