Fixed EC vs VP setting for shared mode LPAR configuration.
In the previously added code, it was observed that when the condition max_proc_units > max_virtual_proc is true, the system assigns the maximum virtual processors as returned by the LPAR configuration. However, an issue arises when the LPAR returns a maximum virtual processor count that is less than the max_proc_units. In such cases, the HMC fails to configure the EC vs VP values according to the virtual processors' thumb rule, which mandates that virtual processors should always be greater than or equal to EC processors.

This patch addresses and resolves the issue.

Signed-off-by: Samir Mulani <samir@linux.vnet.ibm.com>